requirements.txt
streamlit>=1.28.0
audio-recorder-streamlit>=0.0.8
speechrecognition>=3.10.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
python-dotenv>=1.0.0
pydub>=0.25.1
google-generativeai>=0.7.0
